--- TODO.md
@@ -1,23 +1,0 @@
-# TODO
-
-- [ ] **Refactor `LlmClient` Interface:**
-  - [ ] Define a clear, unified `LlmClient` interface that accommodates both Google and OpenAI models for both stateless (`generateContent`) and stateful (`startChat`, `sendMessageStream`) interactions.
-  - [ ] The interface should include methods for `generateContent`, `sendMessageStream`, `generateJson`, `generateEmbedding`, and chat history management.
-- [ ] **Complete `OpenAIClient` Implementation:**
-  - [ ] Implement the `startChat` and `sendMessageStream` methods in `OpenAIClient` to support stateful, streaming conversations.
-  - [ ] Implement `generateJson` for structured data generation.
-  - [ ] Implement `generateEmbedding` for text embeddings.
-  - [ ] Map OpenAI's API responses to the `GenerateContentResponse` and `Turn` types used in the existing codebase.
-  - [ ] Add error handling and retry logic, similar to `GoogleClient`.
-- [ ] **Integrate Langfuse Tracing:**
-  - [ ] Wrap all public methods of both `GoogleClient` and `OpenAIClient` with `langfuseService.trace`. This includes `generateContent`, `sendMessageStream`, `generateJson`, and `generateEmbedding`.
-  - [ ] Ensure that traces capture relevant information, such as the model used, the request, the response, and any errors.
-  - [ ] For streaming methods like `sendMessageStream`, the trace should encompass the entire streaming operation.
-- [ ] **Refactor `getLlmClient`:**
-  - [ ] Update the `getLlmClient` factory function to initialize and return the correct client (`GoogleClient` or `OpenAIClient`) based on the configuration.
-  - [ ] Ensure that the returned client is fully initialized before use.
-- [ ] **Update Configuration:**
-  - [ ] Add configuration options for OpenAI models, including the model name and any other relevant parameters.
-- [ ] **Add Unit and Integration Tests:**
-  - [ ] Write unit tests for the new methods in `OpenAIClient`.
-  - [ ] Write integration tests to verify that both `GoogleClient` and `OpenAIClient` work correctly with the Langfuse integration.
