+++ packages/core/src/core/contentGenerator.ts
@@ -5,23 +5,14 @@
  */
 
 import {
+  CountTokensResponse,
+  GenerateContentResponse,
+  GenerateContentParameters,
+  CountTokensParameters,
+  EmbedContentResponse,
+  EmbedContentParameters,
   GoogleGenAI,
-  CountTokensResponse,
-  GenerateContentResponse,
-  GenerateContentParameters,
-  CountTokensParameters,
-  EmbedContentResponse,
-  EmbedContentParameters,
 } from '@google/genai';
-
-export {
-  CountTokensResponse,
-  GenerateContentResponse,
-  GenerateContentParameters,
-  CountTokensParameters,
-  EmbedContentResponse,
-  EmbedContentParameters,
-};
 import { createCodeAssistContentGenerator } from '../code_assist/codeAssist.js';
 import { DEFAULT_GEMINI_MODEL } from '../config/models.js';
 import { getEffectiveModel } from './modelCheck.js';
@@ -47,7 +38,6 @@ export enum AuthType {
   LOGIN_WITH_GOOGLE_PERSONAL = 'oauth-personal',
   USE_GEMINI = 'gemini-api-key',
   USE_VERTEX_AI = 'vertex-ai',
-  USE_OPENAI = 'openai-api-key',
 }
 
 export type ContentGeneratorConfig = {
@@ -55,18 +45,12 @@ export type ContentGeneratorConfig = {
   apiKey?: string;
   vertexai?: boolean;
   authType?: AuthType | undefined;
-  openaiApiKey?: string;
 };
-
-interface CreateContentGeneratorConfigOptions {
-  getModel?: () => string;
-  openaiApiKey?: string;
-}
 
 export async function createContentGeneratorConfig(
   model: string | undefined,
   authType: AuthType | undefined,
-  config?: CreateContentGeneratorConfigOptions,
+  config?: { getModel?: () => string },
 ): Promise<ContentGeneratorConfig> {
   const geminiApiKey = process.env.GEMINI_API_KEY;
   const googleApiKey = process.env.GOOGLE_API_KEY;
@@ -113,15 +97,6 @@ export async function createContentGeneratorConfig(
     return contentGeneratorConfig;
   }
 
-  if (authType === AuthType.USE_OPENAI) {
-    const effectiveOpenaiApiKey =
-      config?.openaiApiKey ?? process.env.OPENAI_API_KEY;
-    if (effectiveOpenaiApiKey) {
-      contentGeneratorConfig.openaiApiKey = effectiveOpenaiApiKey;
-      return contentGeneratorConfig;
-    }
-  }
-
   return contentGeneratorConfig;
 }
 
@@ -138,8 +113,6 @@ export async function createContentGenerator(
     return createCodeAssistContentGenerator(httpOptions, config.authType);
   }
 
-  let generator: ContentGenerator;
-
   if (
     config.authType === AuthType.USE_GEMINI ||
     config.authType === AuthType.USE_VERTEX_AI
@@ -149,12 +122,11 @@ export async function createContentGenerator(
       vertexai: config.vertexai,
       httpOptions,
     });
-    generator = googleGenAI.models;
-  } else {
-    throw new Error(
-      `Error creating contentGenerator: Unsupported authType: ${config.authType}`,
-    );
+
+    return googleGenAI.models;
   }
 
-  return generator;
+  throw new Error(
+    `Error creating contentGenerator: Unsupported authType: ${config.authType}`,
+  );
 }
