+++ packages/core/src/index.ts
@@ -8,7 +8,7 @@
 export * from './config/config.js';
 
 // Export Core Logic
-
+export * from './core/client.js';
 export * from './core/contentGenerator.js';
 export * from './core/geminiChat.js';
 export * from './core/logger.js';
