--- packages/core/src/telemetry/langfuse.ts
@@ -1,13 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import Langfuse from 'langfuse';
-
-export const langfuse = new Langfuse({
-  publicKey: process.env.LANGFUSE_PUBLIC_KEY,
-  secretKey: process.env.LANGFUSE_SECRET_KEY,
-  baseUrl: process.env.LANGFUSE_BASE_URL ?? 'https://cloud.langfuse.com',
-});
